backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pymongo[srv]==4.7.2
python-dotenv==1.0.1
pydantic==2.7.1
scikit-learn==1.4.2
xgboost==2.0.3
shap==0.45.1
pandas==2.2.2
numpy==1.26.4
langchain==0.2.1
langchain-openai==0.1.7
langchain-community==0.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9